Add files via upload
1.	Discuss the use of throw-away prototype and evolutionary prototype with simple examples. 
Ans:
They are commonly used in early phases of design when a large number of ideas are still being considered. Throw-away prototypes may also be used in late stage design in industries in which products are launched at a low state of refinement. You make a prototype, then when you are done with it, you abandon it. For example, you and your colleagues go to lunch, you come up with an idea of the new product. When you discuss the idea, write down the idea and pull out the original design of the paper napkin. When you return to the office, take out the prototype of the paper napkin and transfer it to your computer. You throw out the napkin. In that case, the napkin is considered a thrown away prototype.
In evolutionary prototyping, the concept of the system will be developed as the project progresses. First of all, we will develop the most visual aspect of the system. You present a portion of the system to the customer and continue to develop prototypes based on the feedback received. At some point, you and the customer agree that the prototype is “good enough” and release the prototype as the final product. It is may be best suited for business systems where developers can interact frequently and informally with end users.